rename woff2 endpoint + improve error handling
